Replaced boundary check
Removed the boundary check function and its calls, replaced it by two if
statements that change the current value to the other border value of
the array if the value is on the border (makes the display=none thing
with the buttons not needed). Oh yeah also changed lowcaseSplit to
lowcaseObject, makes more sense, doesn't it?